Remove test code from the utils module
`createMockRequest` relies on a dev dependency which fails production builds.
To fix this we have split it out to its own module and told `nest build` to
ignore it in `tsconfig.build.json`.
