+++ src/common/utils.ts
@@ -1,5 +1,4 @@
 import { Request } from 'express';
-import { createMock, DeepMocked } from '@golevelup/ts-jest';
 
 export function backLink(req: Request): string {
   const referrer = req.get('Referrer') || '';
@@ -8,19 +7,4 @@ export function backLink(req: Request): string {
   if (referrer.match('^https?://' + host)) {
     return referrer;
   }
-}
-
-export function createMockRequest(
-  referrer: string,
-  host: string,
-): DeepMocked<Request> {
-  return createMock<Request>({
-    get: (header) => {
-      if (header === 'Referrer') {
-        return referrer;
-      } else if (header === 'host') {
-        return host;
-      }
-    },
-  });
 }
